Move OnePath to the Strategies package
