+++ kore/src/Kore/Exec.hs
@@ -55,8 +55,6 @@ import qualified Kore.Internal.Predicate as Predicate
 import           Kore.Internal.TermLike
 import qualified Kore.Logger as Log
 import qualified Kore.ModelChecker.Bounded as Bounded
-import           Kore.OnePath.Verification
-                 ( Claim, defaultStrategy, verify )
 import           Kore.Predicate.Predicate
                  ( makeMultipleOrPredicate, unwrapPredicate )
 import           Kore.Profiler.Data
@@ -88,6 +86,8 @@ import qualified Kore.Step.Simplification.Rule as Rule
 import qualified Kore.Step.Simplification.Simplifier as Simplifier
                  ( create )
 import qualified Kore.Step.Strategy as Strategy
+import           Kore.Strategies.OnePath.Verification
+                 ( Claim, defaultStrategy, verify )
 import           SMT
                  ( MonadSMT, SMT )
