+++ kore/src/Kore/Strategies/OnePath/Verification.hs
@@ -8,7 +8,7 @@ Maintainer  : virgil.serbanuta@runtimeverification.com
 This should be imported qualified.
 -}
 
-module Kore.OnePath.Verification
+module Kore.Strategies.OnePath.Verification
     ( Claim
     , CommonProofState
     , defaultStrategy
